backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pandas==2.1.3
chardet==5.2.0
python-dateutil==2.8.2
xlrd==2.0.1
openpyxl==3.1.2
lxml==4.9.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
